Bug Fix: dovecot SASL failure
The default chroot implementation for smtpd prevents the lib_nss library
from being availble when trying to resolve the `smtpd_sasl_path`
parameter in the format of "inet:<address>:<port". To complicate
matters, `/var/spool/postfix/etc` is empty; does not contain a copy of
`hosts`, `resolv.conf`, or `nsswitch.conf`. While it may be possible to
compensate by adding `hosts` replication to the entrypoint script, or by
exposing the required libraries from `/lib`, it is easier to disable the
jailing.

Notes:
* http://postfix.1071664.n5.nabble.com/No-Invalid-host-lookup-in-smtpd-sasl-path-parameter-td94209.html